pyproject.toml
[project]
name = "django-cognito"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=43.0.3",
    "django-ninja>=1.3.0",
    "django>=5.1.3",
    "pyjwt>=2.10.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.7.4",
]

[tool.ruff]
lint.select = ["I"]

